refactor: changing the popoverPositions to ionPosiitons type
