+++ projects/ion/src/lib/directives/popover/component/popover.component.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/index.scss';
+@import '../../../../styles/index.scss';
 
 $popover-spacing: 14px;
 $popover-width: 299px;
