+++ projects/ion/src/lib/directives/popover/popover.directive.ts
@@ -13,10 +13,10 @@ import {
   ViewContainerRef,
 } from '@angular/core';
 
-import { IonPositionService } from '../position/position.service';
-import { IonPositions } from '../position/types';
-import { getPositionsPopover } from '../utils/popover';
-import { SafeAny } from './../utils/safe-any';
+import { IonPositionService } from '../../position/position.service';
+import { IonPositions } from '../../position/types';
+import { getPositionsPopover } from '../../utils/popover';
+import { SafeAny } from '../../utils/safe-any';
 import { IonPopoverComponent } from './component/popover.component';
 import { PopoverDirectiveProps, PopoverTrigger } from './types';
