+++ projects/ion/src/lib/directives/popover/types.ts
@@ -1,26 +1,11 @@
 import { TemplateRef } from '@angular/core';
 
-import { IonPositions } from '../position/types';
+import { IonPositions } from '../../position/types';
 import { Subject } from 'rxjs';
-import { IonButtonProps } from '../button';
+import { IonButtonProps } from '../../button';
 
 export interface PopoverButtonsProps extends IonButtonProps {
   keepOpenAfterAction?: boolean;
-}
-
-export enum PopoverPosition {
-  TOP_RIGHT = 'topRight',
-  TOP_CENTER = 'topCenter',
-  TOP_LEFT = 'topLeft',
-  RIGHT_TOP = 'rightTop',
-  RIGHT_CENTER = 'rightCenter',
-  RIGHT_BOTTOM = 'rightBottom',
-  LEFT_TOP = 'leftTop',
-  LEFT_CENTER = 'leftCenter',
-  LEFT_BOTTOM = 'leftBottom',
-  BOTTOM_RIGHT = 'bottomRight',
-  BOTTOM_CENTER = 'bottomCenter',
-  BOTTOM_LEFT = 'bottomLeft',
 }
 
 export enum PopoverTrigger {
